chore: delete "createInterceptor" function
